Extract prepare_input to a separate module
